Refactor: recommendation common도메인 보도록 수정
