+++ recommendation/src/main/java/store/_0982/recommendation/RecommendationApplication.java
@@ -2,9 +2,14 @@ package store._0982.recommendation;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @EnableFeignClients
+@EntityScan(basePackages = {
+        "store._0982.recommendation",
+        "store._0982.common.domain"
+})
 @SpringBootApplication
 public class RecommendationApplication {
