+++ recommendation/src/main/java/store/_0982/recommendation/domain/PersonalVectorRepository.java
@@ -1,5 +1,7 @@
 package store._0982.recommendation.domain;
 
+import store._0982.common.domain.vector.PersonalVector;
+
 import java.util.Optional;
 import java.util.UUID;
