+++ recommendation/src/main/java/store/_0982/recommendation/domain/ProductVectorRepository.java
@@ -1,5 +1,7 @@
 package store._0982.recommendation.domain;
 
+import store._0982.common.domain.vector.ProductVector;
+
 import java.util.Optional;
 import java.util.UUID;
