+++ recommendation/src/main/java/store/_0982/recommendation/infrastructure/ProductVectorJpaRepository.java
@@ -1,7 +1,7 @@
 package store._0982.recommendation.infrastructure;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import store._0982.recommendation.domain.ProductVector;
+import store._0982.common.domain.vector.ProductVector;
 
 import java.util.UUID;
