+++ recommendation/src/main/java/store/_0982/recommendation/infrastructure/feign/search/PersonalVectorJpaRepository.java
@@ -1,7 +1,7 @@
 package store._0982.recommendation.infrastructure.feign.search;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import store._0982.recommendation.domain.PersonalVector;
+import store._0982.common.domain.vector.PersonalVector;
 
 import java.util.UUID;
