+++ recommendation/src/main/java/store/_0982/recommendation/infrastructure/feign/search/PersonalVectorRepositoryAdapter.java
@@ -2,7 +2,7 @@ package store._0982.recommendation.infrastructure.feign.search;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import store._0982.recommendation.domain.PersonalVector;
+import store._0982.common.domain.vector.PersonalVector;
 import store._0982.recommendation.domain.PersonalVectorRepository;
 
 import java.util.Optional;
